LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/historical-progress-event.window-expected.txt


PASS <img> does not support ProgressEvent or loadstart/progress/loadend
PASS onloadend is not exposed
